% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.linreg.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients method}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A named vector of coefficients
}
\description{
Coefficients method
}
