\name{Group1Lab4-package}
\alias{Group1Lab4-package}
\alias{Group1Lab4}
\docType{package}
\title{
Group1Lab4:A Package for Solving Problems
}
\description{
\packageDescription{Group1Lab4}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Group1Lab4}
\packageIndices{Group1Lab4}
An overview of how to use the package, including the most important functions
}
\author{
Can Wang and Han Xia

Maintainer: Can Wang (\email{canwa869@student.liu.se})
}
\references{
Literature or other references for background information
}
\keyword{ package }
\seealso{
\code{\link[<pkg>:<pkg>-package]{<pkg>}} 
}
\examples{
# Example of solving a problem:
}
\keyword{package}
