% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Residuals from Linear Regression}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class 'linreg', typically the result of a linear regression model.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
Two ggplot objects are printed: 
1. A plot of residuals vs fitted values.
2. A plot of the square root of standardized residuals vs fitted values.
}
\description{
This function takes a linear regression model and plots the residuals against the fitted values and the standardized residuals against the fitted values.
}
