% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/pred.linreg.R
\name{pred.linreg}
\alias{pred.linreg}
\title{Predicted values method}
\usage{
pred.linreg(object)

pred.linreg(object)
}
\arguments{
\item{object}{An object}
}
\value{
A vector of predicted values

A vector of predicted values
}
\description{
Predicted values method

Predicted values method
}
