% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{pred}
\alias{pred}
\title{Predicted values method}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object of class 'linreg'.}
}
\value{
A vector of predicted values.
}
\description{
Predicted values method
}
