% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{rstandard}
\alias{rstandard}
\title{rstandard method for linreg objects}
\usage{
rstandard(X, ...)
}
\arguments{
\item{X}{An object of class 'linreg'.}

\item{...}{Additional arguments (not used currently).}
}
\value{
A vector of standardized residuals.
}
\description{
rstandard method for linreg objects
}
