% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.linreg.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plot Residuals from Linear Regression}
\usage{
\method{plot}{linreg}(linreg_mod)
}
\arguments{
\item{linreg_mod}{A linear model object.}
}
\value{
Two ggplot objects are printed: 
1. A plot of residuals vs fitted values.
2. A plot of the square root of standardized residuals vs fitted values.
}
\description{
This function takes a linear regression model and plots the residuals
against the fitted values and the standardized residuals against the 
fitted values.
}
