% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/print.linreg.R
\name{print.linreg}
\alias{print.linreg}
\title{print}
\usage{
\method{print}{linreg}(x, ...)

\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
An object

An object
}
\description{
print

print
}
