% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{resid}
\alias{resid}
\title{Residuals method}
\usage{
resid(object, ...)
}
\arguments{
\item{object}{An object of class 'linreg'.}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
A vector of residuals
}
\description{
Residuals method
}
