% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R, R/summary.linreg.R
\name{summary.linreg}
\alias{summary.linreg}
\title{summary}
\usage{
\method{summary}{linreg}(object, ...)

\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{}

\item{...}{Additional arguments to be passed to or from methods.}
}
\value{
summary

summary
}
\description{
summary

summary
}
